chore(jest): Moved ./tests to ./src/tests to satisfy TypeScript in terms of test util/helper imports being inside the project dir
