+++ .npmignore
@@ -3,7 +3,6 @@
 .husky
 .storybook/
 bin/
-tests/
 certificates/
 reports/
 src/
