+++ jest.config.js
@@ -72,10 +72,10 @@ module.exports = {
 
 	// A map from regular expressions to module names that allow to stub out resources with a single module
 	moduleNameMapper: {
-		'~/tests/(.*)$': '<rootDir>/tests/$1',
+		'~/tests/(.*)$': '<rootDir>/src/tests/$1',
 		'~/(.*)$': '<rootDir>/src/$1',
-		'^.+\\.(css|scss)$': '<rootDir>/tests/jest-cssstub.js',
-		'\\.(jpg|jpeg|png|svg)$': '<rootDir>/tests/jest-imagestub.js',
+		'^.+\\.(css|scss)$': '<rootDir>/src/tests/jest-cssstub.js',
+		'\\.(jpg|jpeg|png|svg)$': '<rootDir>/src/tests/jest-imagestub.js',
 		'^react$': '<rootDir>/node_modules/react',
 		'^react-dom$': '<rootDir>/node_modules/react-dom',
 	},
@@ -114,7 +114,7 @@ module.exports = {
 	// rootDir: null,
 
 	// A list of paths to directories that Jest should use to search for files in
-	roots: ['<rootDir>/src', '<rootDir>/tests'],
+	roots: ['<rootDir>/src'],
 
 	// Allows you to use a custom runner instead of Jest's default test runner
 	// runner: "jest-runner",
@@ -123,7 +123,7 @@ module.exports = {
 	setupFiles: [],
 
 	// A list of paths to modules that run some code to configure or set up the testing framework before each test
-	setupFilesAfterEnv: ['./tests/jest-setup.ts'],
+	setupFilesAfterEnv: ['./src/tests/jest-setup.ts'],
 
 	// A list of paths to snapshot serializer modules Jest should use for snapshot testing
 	snapshotSerializers: [],
@@ -142,7 +142,7 @@ module.exports = {
 	//   "**/__tests__/**/*.[jt]s?(x)",
 	//   "**/?(*.)+(spec|test).[tj]s?(x)"
 	// ],
-	testMatch: ['**/src/**/*.test.(js|jsx|ts|tsx)', '**/tests/**/*.test.(js|jsx|ts|tsx)'],
+	testMatch: ['**/src/**/*.test.(js|jsx|ts|tsx)'],
 
 	// An array of regexp pattern strings that are matched against all test paths, matched tests are skipped
 	// testPathIgnorePatterns: [
