+++ package.json
@@ -18,13 +18,13 @@
     "test": "jest",
     "test:watch": "jest --watch",
     "lint": "npm-run-all --parallel lint:js lint:css lint:types",
-    "lint:js": "eslint --ignore-path .eslintignore -c .eslintrc.js \"src/**/*.{ts,tsx,js,jsx}\" \"tests/**/*.{ts,tsx,js,jsx}\"",
+    "lint:js": "eslint --ignore-path .eslintignore -c .eslintrc.js \"src/**/*.{ts,tsx,js,jsx}\" \"src/tests/**/*.{ts,tsx,js,jsx}\"",
     "lint:css": "stylelint --config-basedir ./ --ignore-path .stylelintignore --config .stylelintrc.js \"src/**/*.{tsx,ts}\"",
     "lint:types": "tsc --noEmit",
     "fix": "npm-run-all --parallel fix:js fix:css",
-    "fix:js": "eslint --fix --ignore-path .eslintignore -c .eslintrc.js \"src/**/*.{ts,tsx,js,jsx}\" \"tests/**/*.{ts,tsx,js,jsx}\"",
+    "fix:js": "eslint --fix --ignore-path .eslintignore -c .eslintrc.js \"src/**/*.{ts,tsx,js,jsx}\" \"src/tests/**/*.{ts,tsx,js,jsx}\"",
     "ci:test": "jest --ci --reporters=default",
-    "ci:eslint": "eslint --ignore-path .eslintignore -c .eslintrc.js \"src/**/*.{ts,tsx,js,jsx}\" \"tests/**/*.{ts,tsx,js,jsx}\" --format junit --output-file reports/eslint.xml",
+    "ci:eslint": "eslint --ignore-path .eslintignore -c .eslintrc.js \"src/**/*.{ts,tsx,js,jsx}\" \"src/tests/**/*.{ts,tsx,js,jsx}\" --format junit --output-file reports/eslint.xml",
     "ci:stylelint": "npm run stylelint -- --quiet --custom-formatter node_modules/stylelint-junit-formatter > reports/stylelint.xml",
     "ci:lint": "npm-run-all --parallel --silent --continue-on-error ci:eslint ci:stylelint",
     "storybook": "start-storybook -p 6006",
