+++ src/components/Form/__mocks__/Form.tsx
@@ -1,4 +1,4 @@
-import { createMockComponent } from '~/tests/helpers/createMockComponent';
+import { createMockComponent } from '../../../tests/helpers/createMockComponent';
 
 const Form = createMockComponent('Form', 'form');
