+++ src/components/MetadataForm/__mocks__/MetadataForm.tsx
@@ -1,4 +1,4 @@
-import { createMockComponent } from '~/tests/helpers/createMockComponent';
+import { createMockComponent } from '../../../tests/helpers/createMockComponent';
 
 const MetadataForm = createMockComponent('MetadataForm', 'metadata-form');
