+++ src/features/single-file/components/AddDocumentButton/AddDocumentButton.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import AddDocumentButton, { Props } from './AddDocumentButton';
-import renderWithProviders from '~/tests/utils/withProviders';
+import renderWithProviders from '../../../../tests/utils/withProviders';
 import Step1, { SupportedHTTPMethods } from '../Wizard/Step1';
 import { MetadataExample } from 'src/types';
 import { asset } from '../Wizard/__stubs__';
