+++ src/features/single-file/components/Wizard/__mocks__/Step1.tsx
@@ -1,4 +1,4 @@
-import { createMockComponent } from '~/tests/helpers/createMockComponent';
+import { createMockComponent } from '../../../../../tests/helpers/createMockComponent';
 
 const Step1 = createMockComponent('Step1', 'step-1');
