+++ src/tests/jest-imagestub.js
