+++ src/tests/utils/withTheme.tsx
@@ -3,7 +3,7 @@ import { RenderOptions, render } from '@testing-library/react';
 import { muiTheme } from '@amsterdam/bmi-component-library';
 import { GlobalStyle, ThemeProvider } from '@amsterdam/asc-ui';
 import { ThemeProvider as MUIThemeProvider } from '@material-ui/core/styles';
-import theme from '../../src/theme';
+import theme from '../../theme';
 
 function renderWithTheme(ui: React.ReactElement, options?: Omit<RenderOptions, 'queries'>) {
 	const ThemeProviders: React.FC = ({ children }) => (
